requirements.txt
gradio==5.0.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
torch>=2.0.0
transformers>=4.41.0
langchain==0.2.5
ollama==0.2.1

